# cpu: pooling: improve nchw_pooling performance