## Description

Small changes that improve performance of simple_nchw pooling implementation (nchw_pooling_fwd_t/nchw_pooling_bwd_t) in the following cases: max pooling in forward direction, average pooling in backward direction.

Both versions of the library were built with g++ 8.3.0. Tests run on JF5300-B11A332T.

The script used for testing:
```
export KMP_AFFINITY=granularity=fine,compact,1,0
export OMP_NUM_THREADS=56
export LD_LIBRARY_PATH=./:$LD_LIBRARY_PATH

libiomp5_loc=/home/sdp/akharito/19.0/20210329_000000/lnx32e/compiler/lib/intel64_lin
export LD_PRELOAD=${libiomp5_loc}/libiomp5.so

numactl --physcpubind=0-55 --membind=0 ./benchdnn -v5 --mode=p --pool "$@"
```
## Test results

| script arguments (56 threads, mb=500) | original (ms) | fixed (ms) | ratio |
| --- | --- | --- | --- |
| --dt=f32:f32 --alg=avg_np --dir=BWD_D mb500ic16id60ih60iw60od31oh31ow31kd4kh2kw3sd2sh2sw2pd1ph1pw1 | 187.384 |	138.586 | 0.74 |
| --dt=f32:f32 --alg=max --dir=FWD_D mb500ic16id60ih60iw60od31oh31ow31kd4kh2kw3sd2sh2sw2pd1ph1pw1 | 576.944 |	436.926 | 0.76 |
| --dt=f32:f32 --alg=max --dir=FWD_I mb500ic16id60ih60iw60od31oh31ow31kd4kh2kw3sd2sh2sw2pd1ph1pw1 | 546.457	| 428.587 | 0.78 |
|  |  |  |  |
| --skip-impl=jit --dt=f16:f16 --alg=avg_np --dir=BWD_D mb500ic16id60ih60iw60od31oh31ow31kd4kh2kw3sd2sh2sw2pd1ph1pw1 | 163.594 | 117.859 | 0.72 |
| --skip-impl=jit --dt=f16:f16 --alg=max --dir=FWD_D mb500ic16id60ih60iw60od31oh31ow31kd4kh2kw3sd2sh2sw2pd1ph1pw1 | 691.488 | 560.89 | 0.81 |
| --skip-impl=jit --dt=f16:f16 --alg=max --dir=FWD_I mb500ic16id60ih60iw60od31oh31ow31kd4kh2kw3sd2sh2sw2pd1ph1pw1 | 665.588 | 547.249 | 0.82 |
|  |  |  |  |
| --skip-impl=jit --dt=bf16:bf16 --alg=avg_np --dir=BWD_D mb500ic16id60ih60iw60od31oh31ow31kd4kh2kw3sd2sh2sw2pd1ph1pw1 | 159.371 | 121.846 | 0.77 |
| --skip-impl=jit --dt=bf16:bf16 --alg=max --dir=FWD_D mb500ic16id60ih60iw60od31oh31ow31kd4kh2kw3sd2sh2sw2pd1ph1pw1 | 749.277 | 629.006 | 0.84 |
| --skip-impl=jit --dt=bf16:bf16 --alg=max --dir=FWD_I mb500ic16id60ih60iw60od31oh31ow31kd4kh2kw3sd2sh2sw2pd1ph1pw1 | 716.403 | 607.367 | 0.85 |
|  |  |  |  |
| --skip-impl=jit --dt=f8_e5m2:f8_e5m2 --alg=max --dir=FWD_D mb500ic16id60ih60iw60od31oh31ow31kd4kh2kw3sd2sh2sw2pd1ph1pw1 | 867.921 | 750.832 | 0.87 |
| --skip-impl=jit --dt=f8_e5m2:f8_e5m2 --alg=max --dir=FWD_I mb500ic16id60ih60iw60od31oh31ow31kd4kh2kw3sd2sh2sw2pd1ph1pw1 | 842.758 | 733.6 | 0.87 |
| --skip-impl=jit --dt=f8_e4m3:f8_e4m3 --alg=max --dir=FWD_D mb500ic16id60ih60iw60od31oh31ow31kd4kh2kw3sd2sh2sw2pd1ph1pw1 | 1148.24 | 1006.06 | 0.88 |
| --dt=f32:f32 --alg=max --dir=FWD_I mb100ic64_id64od32_kd2sd2n"3d_unet:max_pool1" | 259.428 | 214.501 | 0.83 |
| --skip-impl=jit --dt=f32:f32 --alg=max --dir=FWD_I mb100ic128_id28od14_kd2sd2n"3d_unet:max_pool2" | 43.2533 | 37.2559 | 0.86 |
| --dt=f32:f32 --alg=max --dir=FWD_I mb100ic128_id56od28_kd2sd2n"3d_unet:max_pool2" | 353.548 | 294.497 | 0.83 |


| script arguments (1 thread, mb=100) | original (ms) | fixed (ms) | ratio |
| --- | --- | --- | --- |
| --dt=f32:f32 --alg=avg_np --dir=BWD_D mb100ic16id60ih60iw60od31oh31ow31kd4kh2kw3sd2sh2sw2pd1ph1pw1 | 1256.2 | 916.63 | 0.73 |
| --dt=f32:f32 --alg=max --dir=FWD_D mb100ic16id60ih60iw60od31oh31ow31kd4kh2kw3sd2sh2sw2pd1ph1pw1 | 4418.59 | 3315.03 | 0.75 |
| --dt=f32:f32 --alg=max --dir=FWD_I mb100ic16id60ih60iw60od31oh31ow31kd4kh2kw3sd2sh2sw2pd1ph1pw1 | 4224.97 | 	3235.24 | 0.77 |
|  |  |  |  |
| --skip-impl=jit --dt=f16:f16 --alg=avg_np --dir=BWD_D mb100ic16id60ih60iw60od31oh31ow31kd4kh2kw3sd2sh2sw2pd1ph1pw1 | 1141.21 | 847.603 | 0.74 |
| --skip-impl=jit --dt=f16:f16 --alg=max --dir=FWD_D mb100ic16id60ih60iw60od31oh31ow31kd4kh2kw3sd2sh2sw2pd1ph1pw1 | 5139.02 | 4137.88 | 0.81 |
| --skip-impl=jit --dt=f16:f16 --alg=max --dir=FWD_I mb100ic16id60ih60iw60od31oh31ow31kd4kh2kw3sd2sh2sw2pd1ph1pw1 | 4921.83 | 4052.5 | 0.82 |